Move the logic to register test adapters to the 0.6.x-only Settings.
Now that the 1.x Settings do not create their own TestAdapters,
they do not need the infrastructure. We move it from the settings
that apply to both 0.6.x and 1.x to the 0.6.x-only Settings. This
allows to remove the mentions to `TestAdapter` in the `compat`
objects.
